Move `EdgeRuntime` to `EdgeVM`
