+++ docs/pages/packages/_meta.json
@@ -1,6 +1,5 @@
 {
   "cookies": "@edge-runtime/cookies",
-  "feature-detector": "@edge-runtime/feature-detector",
   "format": "@edge-runtime/format",
   "jest-environment": "@edge-runtime/jest-environment",
   "jest-expect": "@edge-runtime/jest-expect",
