--- docs/pages/packages/feature-detector.mdx
@@ -1,66 +1,0 @@
-import { Callout } from 'nextra-theme-docs'
-import { Tabs, Tab } from '../../components/tabs'
-
-# Edge Runtime Code analysis utilities
-
-The **@edge-runtime/feature-detector** package contains utilities to analyze code running on the edge and its used low-level APIs.
-
-It leverages the excellent [ts-morph](https://ts-morph.com/) package, which is a wrapper around TypeScript compiler API to navigate its Abstract Syntax Tree (AST).
-
-It can analyse JavaScript and TypeScript code.
-
-<Callout type='warning' emoji='⚠️'>
-  Please note this is an alpha version.
-</Callout>
-
-## Installation
-
-<Tabs items={['npm', 'yarn', 'pnpm']} storageKey='selected-pkg-manager'>
-  <Tab>```sh npm install @edge-runtime/feature-detector --save ```</Tab>
-  <Tab>```sh yarn add @edge-runtime/feature-detector ```</Tab>
-  <Tab>```sh pnpm install @edge-runtime/feature-detector --save ```</Tab>
-</Tabs>
-
-This package includes built-in TypeScript support.
-
-## Usage
-
-Here is a test file. It's in JavaScript, but TypeScript is supported as well.
-
-```js
-// test.js
-if (typeof process !== 'undefined') {
-  console.log('in node', __dirname)
-} else if (typeof window !== 'undefined') {
-  console.log('in a browser', document)
-}
-setTimeout(() => console.log('setTimeout is not a global, nor console'), 0)
-
-export default function () {
-  return Response.json({ message: 'hello world!' })
-}
-```
-
-In the code snippet bellow, we're checking:
-
-1. whether this file is a function matching the edge signature
-2. globals used, that are not provided by the Edge Runtime
-
-```ts
-import { hasEdgeSignature, findGlobals } from '@edge-runtime/feature-detector'
-
-const sourceFilePath = './test.js' // could be TypeScript as well. Must be in current working directory
-
-// 1
-if (hasEdgeSignature(sourceFilePath)) {
-  console.log(`${sourcefilePath} can run on the edge`)
-}
-// 2
-console.log(findGlobals(sourceFilePath)) // ['process', '__dirname', 'window', 'document']
-```
-
-## API
-
-<Callout type='warning' emoji='⚠️'>
-  Work in progress
-</Callout>
