--- packages/feature-detector/CHANGELOG.md
@@ -1,23 +1,0 @@
-# @edge-runtime/feature-detector
-
-## 1.0.2
-
-### Patch Changes
-
-- Use valid SPDX license expression ([#276](https://github.com/vercel/edge-runtime/pull/276))
-
-## 1.0.1
-
-### Patch Changes
-
-- build: update dependencies ([#271](https://github.com/vercel/edge-runtime/pull/271))
-
-## 1.0.0
-
-### Minor Changes
-
-- Introduces findGlobals() utility ([#234](https://github.com/vercel/edge-runtime/pull/234))
-
-### Patch Changes
-
-- introduce hasEdgeSignature detection utility ([#226](https://github.com/vercel/edge-runtime/pull/226))
