--- packages/feature-detector/README.md
@@ -1,38 +1,0 @@
-<div align="center">
-  <br>
-  <img src="https://edge-runtime.vercel.app/og-image.png" alt="edge-runtime logo">
-  <br>
-  <br>
-  <p align="center"><strong>@edge-runtime/feature-detector</strong>: A set of utilities to analyze code running on the edge and its used low-level APIs.</p>
-  <p align="center">See <a href="https://edge-runtime.vercel.app/packages/feature-detector" target='_blank' rel='noopener noreferrer'>@edge-runtime/feature-detector</a> section in our <a href="https://edge-runtime.vercel.app/" target='_blank' rel='noopener noreferrer'>website</a> for more information.</p>
-  <br>
-</div>
-
-> **Note**: This is an alpha version.
-
-## Install
-
-Using npm:
-
-```sh
-npm install @edge-runtime/feature-detector --save
-```
-
-or using yarn:
-
-```sh
-yarn add @edge-runtime/feature-detector --dev
-```
-
-or using pnpm:
-
-```sh
-pnpm install @edge-runtime/feature-detector --save
-```
-
-## License
-
-**@edge-runtime/feature-detector** © [Vercel](https://vercel.com), released under the [MPLv2](https://github.com/vercel/edge-runtime/blob/main/LICENSE.md) License.<br>
-Authored and maintained by [Vercel](https://vercel.com) with help from [contributors](https://github.com/vercel/edge-runtime/contributors).
-
-> [vercel.com](https://vercel.com) · GitHub [Vercel](https://github.com/vercel) · Twitter [@vercel](https://twitter.com/vercel)
