--- packages/feature-detector/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@edge-runtime/feature-detector",
-  "description": "A set of utilities to analyze code running on the edge and its used low-level APIs",
-  "homepage": "https://edge-runtime.vercel.app/packages/feature-detector",
-  "version": "1.0.2",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "repository": {
-    "directory": "packages/feature-detector",
-    "type": "git",
-    "url": "git+https://github.com/vercel/edge-runtime.git"
-  },
-  "bugs": {
-    "url": "https://github.com/vercel/edge-runtime/issues"
-  },
-  "keywords": [
-    "detector",
-    "edge",
-    "edge-runtime",
-    "feature",
-    "functions",
-    "runtime",
-    "standard",
-    "web"
-  ],
-  "dependencies": {
-    "ts-morph": "17.0.1"
-  },
-  "devDependencies": {
-    "axios": "1.3.4",
-    "jest-text-transformer": "1.0.4",
-    "next": "13",
-    "react": "18",
-    "react-dom": "18",
-    "tsup": "6"
-  },
-  "engines": {
-    "node": ">=14"
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsup",
-    "clean:build": "rm -rf dist",
-    "prebuild": "pnpm run clean:build",
-    "test": "jest"
-  },
-  "license": "MPL-2.0",
-  "publishConfig": {
-    "access": "public"
-  },
-  "types": "dist/index.d.ts"
-}
