--- packages/feature-detector/src/index.ts
@@ -1,2 +1,0 @@
-export * from './find-globals'
-export * from './has-edge-signature'
