--- packages/feature-detector/src/utils/type-definition.txt
@@ -1,167 +1,0 @@
-/* We can not load default library types because it will bring too many of them, which the edge-runtime does not support */
-/// <reference no-default-lib="true"/>
-
-/* We load the minimal ES version supported by the edge-runtime (aka node@14) */
-/// <reference lib="es2019" />
-
-/*
-Then we only need to declare edge-supported classes, functions and variables that are not part of ES standard, and supported by the runtime.
-We don't need to fully declare them, except for the their static methods (because they produce instances of the var, like the new operator).
-The simple declaration is enough for ts-morph to assign a Symbol to identifiers, which makes the distinctions between runtime builtins and undefined globals.
-*/
-interface AbortController {}
-declare var AbortController: {
-  prototype: AbortController;
-  new(): AbortController;
-};
-
-interface AbortSignal {}
-declare var AbortSignal: {
-  prototype: AbortSignal;
-  new(): AbortSignal;
-  timeout(): AbortSignal;
-  abort(): AbortSignal;
-};
-
-interface Blob {}
-declare var Blob: {
-  prototype: Blob;
-  new(): Blob;
-};
-
-interface Crypto {}
-declare var Crypto: {
-  prototype: Crypto;
-  new(): Crypto;
-};
-
-interface CryptoKey {}
-declare var CryptoKey: {
-  prototype: CryptoKey;
-  new(): CryptoKey;
-};
-
-interface DOMException {}
-declare var DOMException: {
-  prototype: DOMException;
-  new(): DOMException;
-};
-
-interface FetchEvent {}
-declare var FetchEvent: {
-  prototype: FetchEvent;
-  new(): FetchEvent;
-};
-
-interface File {}
-declare var File: {
-  prototype: File;
-  new(): File;
-};
-
-interface FormData {}
-declare var FormData: {
-  prototype: FormData;
-  new(): FormData;
-};
-
-interface Headers {}
-declare var Headers: {
-  prototype: Headers;
-  new(): Headers;
-};
-
-interface SubtleCrypto {}
-declare var SubtleCrypto: {
-  prototype: SubtleCrypto;
-  new(): SubtleCrypto;
-};
-
-interface TextEncoder {}
-declare var TextEncoder: {
-  prototype: TextEncoder;
-  new(): TextEncoder;
-};
-
-interface TextDecoder {}
-declare var TextDecoder: {
-  prototype: TextDecoder;
-  new(): TextDecoder;
-};
-
-interface ReadableStream {}
-declare var ReadableStream: {
-  prototype: ReadableStream;
-  new(): ReadableStream;
-};
-
-interface ReadableStreamBYOBReader {}
-declare var ReadableStreamBYOBReader: {
-  prototype: ReadableStreamBYOBReader;
-  new(): ReadableStreamBYOBReader;
-};
-
-interface ReadableStreamDefaultReader {}
-declare var ReadableStreamDefaultReader: {
-  prototype: ReadableStreamDefaultReader;
-  new(): ReadableStreamDefaultReader;
-};
-
-interface Request {}
-declare var Request: {
-  prototype: Request;
-  new(): Request;
-};
-
-interface Response {}
-declare var Response: {
-  prototype: Response;
-  new(): Response;
-  error(): Response;
-  redirect(): Response;
-  json(): Response;
-};
-
-interface TransformStream {}
-declare var TransformStream: {
-  prototype: TransformStream;
-  new(): TransformStream;
-};
-
-interface URL {}
-declare var URL: {
-  prototype: URL;
-  new(): URL;
-};
-
-interface URLSearchParams {}
-declare var URLSearchParams: {
-  prototype: URLSearchParams;
-  new(): URLSearchParams;
-};
-
-interface URLPattern {}
-declare var URLPattern: {
-  prototype: URLPattern;
-  new(): URLPattern;
-};
-
-interface WritableStream {}
-declare var WritableStream: {
-  prototype: WritableStream;
-  new(): WritableStream;
-};
-
-interface WritableStreamDefaultWriter {}
-declare var WritableStreamDefaultWriter: {
-  prototype: WritableStreamDefaultWriter;
-  new(): WritableStreamDefaultWriter;
-};
-
-declare var atob: (encoded: string) => string;
-declare var btoa: (str: string) => string;
-declare var console: {};
-declare var crypto: Crypto;
-declare var fetch: () => Promise<Response>;
-declare var setTimeout: () => number;
-declare var structuredClone: <T>(any: T) => T;
