--- packages/feature-detector/test/fixtures/with-axios.ts
@@ -1,5 +1,0 @@
-import axios from 'axios'
-
-export default async function () {
-  return axios.get('https://example.vercel.sh').then((res) => res.data)
-}
