--- packages/feature-detector/tsconfig.prod.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src/**/*.ts"],
-  "types": ["node", "web"]
-}
