--- packages/feature-detector/tsup.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from 'tsup'
-
-export default defineConfig({
-  dts: { resolve: true },
-  entry: ['./src/index.ts'],
-  format: ['cjs', 'esm'],
-  tsconfig: './tsconfig.prod.json',
-})
