+++ packages/jest-environment/package.json
@@ -23,7 +23,7 @@
     "web"
   ],
   "dependencies": {
-    "@edge-runtime/vm": "workspace:2.1.2",
+    "@edge-runtime/vm": "workspace:*",
     "@jest/environment": "28.1.3",
     "@jest/fake-timers": "28.1.3",
     "@jest/types": "28.1.3",
