+++ packages/ponyfill/package.json
@@ -27,9 +27,9 @@
     "web"
   ],
   "devDependencies": {
-    "@edge-runtime/jest-environment": "workspace:2.0.8",
-    "@edge-runtime/primitives": "workspace:2.1.2",
-    "@edge-runtime/vm": "workspace:2.1.2",
+    "@edge-runtime/jest-environment": "workspace:*",
+    "@edge-runtime/primitives": "workspace:*",
+    "@edge-runtime/vm": "workspace:*",
     "acorn": "8.8.2",
     "acorn-loose": "8.3.0",
     "acorn-walk": "8.2.0"
