+++ packages/runtime/src/types.ts
@@ -18,19 +18,3 @@ export interface Logger {
 export interface NodeHeaders {
   [header: string]: string | string[] | undefined
 }
-
-export interface DispatchFetch {
-  (input: string, init?: RequestInit): Promise<
-    Response & {
-      waitUntil: () => Promise<any>
-    }
-  >
-}
-
-export interface RejectionHandler {
-  (reason?: {} | null, promise?: Promise<any>): void
-}
-
-export interface ErrorHandler {
-  (error?: {} | null): void
-}
