--- packages/runtime/tests/fixtures/legit-uncaught-exception.ts
@@ -1,4 +1,0 @@
-import { EdgeRuntime } from '../../src'
-
-new EdgeRuntime()
-throw new Error('intentional break')
