--- packages/runtime/tests/fixtures/legit-unhandled-rejection.ts
@@ -1,4 +1,0 @@
-import { EdgeRuntime } from '../../src'
-
-new EdgeRuntime()
-Promise.reject(new Error('intentional break'))
