+++ packages/vm/tests/fixtures/cjs-module.js
