+++ packages/vm/tests/fixtures/legit-uncaught-exception.ts
@@ -1,0 +1,4 @@
+import { EdgeVM } from '../../src'
+
+new EdgeVM()
+throw new Error('intentional break')
