+++ packages/vm/tests/fixtures/legit-unhandled-rejection.ts
@@ -1,0 +1,4 @@
+import { EdgeVM } from '../../src/edge-vm'
+
+new EdgeVM()
+Promise.reject(new Error('intentional break'))
