+++ packages/vm/tests/fixtures/uncaught-exception.ts
@@ -1,8 +1,8 @@
-import { EdgeRuntime } from '../../src'
+import { EdgeVM } from '../../src/edge-vm'
 import assert from 'assert'
 
 function main() {
-  const runtime = new EdgeRuntime()
+  const runtime = new EdgeVM()
   runtime.context.handleError = (error: Error) => {
     assert.strictEqual(error?.message, 'expected error')
     console.log('TEST PASSED!')
@@ -12,7 +12,7 @@ function main() {
     addEventListener('error', (error) => {
       globalThis.handleError(error)
     })
-    
+
     throw new Error('expected error')
   `)
 }
